문제코드 수정
잘못된 코드를 수정하고 문제점 서술

+++ 07.Special Ball Throw-1/code.c
@@ -2,10 +2,10 @@
 
 int main(void)
 {
-    int a;
+    double a;
     
-    scanf("%d", &a);
-    if (a >= 50 && a <= 60) {
+    scanf("%f", &a);  // 정수가 아닌 실수임을 고려하지 못함
+    if (a >= 50.0 && a <= 60.0) {
         printf("win");
     } else {
         printf("lose");
