+++ 09.Special ball throw-3/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
 int main() {
-    double a;
-    scanf("%lf", &a);
-    
-    if (a >= 50 && a <= 70 || a % 6 == 0) {
+    int a;
+    scanf("%d", &a);
+    // and 부분에 괄호를 쳐서 먼저 계산이 되게 해야함.
+    if ((a >= 50 && a <= 70) || a % 6 == 0) {
         printf("win");
     } else {
         printf("lose");
